device_gpu/requirements.txt
transformers==4.45.1
pillow==10.2.0
opencv-python==4.10.0.84
scikit-image==0.24.0
numpy==2.1.1